+++ src/index.js
@@ -1,10 +1,11 @@
 /**
  * @module  nodeUtils
  */
+
 module.exports = {
     ConvName: require('./utils/ConvName'),
     FastPath: require('./utils/FastPath'),
     FastFs: require('./utils/FastFs'),
     Notice: require('./utils/Notice'),
-    getDeepPreset: require('./utils/getDeepPreset'),
+    PresetUtils: require('./utils/PresetUtils'),
 };