+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.0.2](https://gitee.com/agile-development-system/node-utils/compare/v1.0.1...v1.0.2) (2021-04-25)
+
+
+### Features
+
+* **all:** 优化jsdoc注释,以module命名模块,新增getPresetConfig函数,新增CHANGELOG相关配置 ([7805d92](https://gitee.com/agile-development-system/node-utils/commits/7805d924a06ad4a4490eed323414b6c90784bf14))
+* **presetutils:** 修改preset相关处理工具,新增merge ([3a285fb](https://gitee.com/agile-development-system/node-utils/commits/3a285fbac014a48230dbb96c038897d98a9952f6))
+
+
+
 ## [1.0.2](https://gitee.com/agile-development-system/node-utils/compare/v1.0.1...v1.0.2) (2021-04-22)